pyproject.toml
[tool.poetry]
name = "func-ai"
version = "0.0.9"
description = "AI Functional Catalog - OpenAI functions on steriods"
authors = ["Trayan Azarov <trayan.azarov@amikos.tech>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "func_ai" }]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/amikos-tech/func-ai/issues"
"Homepage" = "https://github.com/amikos-tech/func-ai/"
"Source" = "https://github.com/amikos-tech/func-ai/"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
requests = "^2.31.0"
chromadb = "^0.3.26"
python-ulid = "^1.1.0"
fastapi = "^0.99.1"
pyyaml = "^6.0"
jsonschema = "^4.17.3"
python-dotenv = "^1.0.0"
tenacity = "^8.2.2"
jinja2 = "^3.1.2"
gradio = "^3.35.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-env = "^0.8.2"
pytest-html = "^3.2.0"
pytest-metadata = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
